+++ dkg-gadget/src/lib.rs
@@ -33,11 +33,11 @@ mod gossip;
 mod keyring;
 mod keystore;
 pub mod messages;
-mod meta_async_protocol;
 mod metrics;
 mod persistence;
 mod proposal;
 pub mod storage;
+mod meta_async_rounds;
 mod types;
 mod utils;
 mod worker;
