+++ dkg-gadget/src/metrics.rs
@@ -18,11 +18,7 @@ use prometheus::{register, Gauge, PrometheusError, Registry, U64};
 /// DKG metrics exposed through Prometheus
 pub(crate) struct Metrics {
 	/// Current active validator set id
-	pub dkg_validator_set_id: Gauge<U64>,
-	/// Total number of votes sent by this node
-	pub dkg_votes_sent: Gauge<U64>,
-	/// Most recent concluded voting round
-	pub dkg_round_concluded: Gauge<U64>,
+	pub dkg_validator_set_id: Gauge<U64>
 }
 
 impl Metrics {
@@ -30,14 +26,6 @@ impl Metrics {
 		Ok(Self {
 			dkg_validator_set_id: register(
 				Gauge::new("dkg_validator_set_id", "Current DKG active validator set id.")?,
-				registry,
-			)?,
-			dkg_votes_sent: register(
-				Gauge::new("dkg_votes_sent", "Number of votes sent by this node")?,
-				registry,
-			)?,
-			dkg_round_concluded: register(
-				Gauge::new("dkg_round_concluded", "Voting round, that has been concluded")?,
 				registry,
 			)?,
 		})
