+++ dkg-primitives/src/lib.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 //
 pub mod keys;
-pub mod rounds;
 pub mod types;
 pub mod utils;
