+++ dkg-primitives/src/types.rs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-use crate::rounds::LocalKey;
 use codec::{Decode, Encode};
 use curv::elliptic::curves::{Point, Scalar, Secp256k1};
 use dkg_runtime_primitives::{crypto::AuthorityId, MisbehaviourType};
 use std::fmt;
+use multi_party_ecdsa::protocols::multi_party_ecdsa::gg_2020::state_machine::keygen::LocalKey;
 
 pub type FE = Scalar<Secp256k1>;
 pub type GE = Point<Secp256k1>;
